# Conflicts:
#   backend/src/controllers/PostController.js

+++ backend/src/controllers/PostController.js
@@ -60,12 +60,28 @@ export const updatePost = async (req, res) => {
 // Delete los Post de un User
 export const deletePost = async (req, res) => {
   try {
-    const { postId } = req.body
-    const post = await Post.deleteOne({ _id: postId })
-    if (post.deletedCount === 0)
-      return res.status(404).json({ error: 'No se ha podido borrar el post' })
-    res.json(post)
+    const postId = req.params.id;        // ID del post desde la URL
+    const userId = req.user.id;          // ID del usuario logueado (del token)
+
+    //Buscar el post en la BBDD
+    const post = await Post.findById(postId);
+
+    if (!post) {
+      return res.status(404).json({ error: 'Post no encontrado' });
+    }
+
+    // Comprovar que el usuario logueado sea el que creo ese post
+    if (post.user.toString() !== userId) {
+      return res.status(403).json({ error: 'No tienes permiso para borrar este post' });
+    }
+
+    // Borramos post
+    await Post.deleteOne({ _id: postId });
+
+    res.status(200).json({ message: 'Post eliminado correctamente' });
+
   } catch (err) {
-    res.status(500).json({ error: err.message })
+    console.error(err);
+    res.status(500).json({ error: err.message });
   }
 }
